[gen/wiki/oar] Move network_address from hiearchy to misc properties
Because 'network_address' is now blocked by the resource hierarchy
safeguard.

I checked: no user used it in the oarsub `-l` part til now, and no
documentation use it either.

'host' must be used, or 'nodes'.

'network_address' can still be used as a property filter in oarsub (but
using 'host' is shorted to write for the same result).