[gen/wiki] Fix wiki links to exotic clusters by simplifying hardware page headings
The page heading for each cluster currently contains the queue and the job
type, for instance:

== pyxis (exotic job type) ==
== graoully (production queue) ==
== drac (testing queue, exotic job type) ==

It means that anchors links are horrible, because they are automatically
generated by Mediawiki.  For instance:

  https://www.grid5000.fr/w/Nancy:Hardware#graoully_.28production_queue.29

We try to generate these horrible links in several places in the wiki
generation code, which is difficult.  In addition, it fails to take into
account the exotic job type, so all links to exotic clusters are currently
broken.

Fix this by only keeping the cluster name in the heading: this way, we
don't need to guess the horrible link anchors generated by Mediawiki.
The missing information (queue and job type i.e. "access condition") is
moved to the table presenting the cluster characteristics.