[gen/wiki/site_hardware] do not show by_path for disk
With some HPE equipment, the by_path contains an unique id for each disk,
so the nodes from a cluster cannot be grouped in the site_hardware page.
This commit remove the by_path from this page.
The path based on id of the disk is still given and usable (eg: /dev/disk1), and
the by_path is still available from the API.