[lib] [bindg5k] Stop updating DNS serial
Currently, we auto-generate a serial number for all our DNS zones.  This
generates a lot of merge conflicts when dealing with multiple git
branches.

This is actually not needed for internal zones, because we don't have any
DNS slaves (all DNS servers are master and we rely on Puppet to update
zone content).

However, this is still required for the external zones, because
dns.inria.fr is a DNS slave and it relies on the zone serial to update its
local copy of the zones.